[MBL 6908] Support file upload quiz question type (#228)
* Add DocumentMenuViewModel helper

* Refactor managed object observer lifetime stuff

It is now the responsibility of the caller to specify the lifetime.
This is so that the caller has flexibility about the lifetime.

* Add a callback-based file uploader

* Support file upload quiz question type

refs MBL-6908

* Better error handling

* Add action to remove file answer

* Don't assume textLabel is going to be there

* Fix choosing file on ipad

* Fix choosing document picker files

* Fix video uploads

Makes asset uploads smarter and easier to generate all around.

* Use video file name for videos taken on the fly

The name must include the extension otherwise chrome downloads it
as a plain text file.

* Fix displaying upload errors

* Preserve original asset file sizes

* Fix file uploads after being logged out

For whatever reason, the response we were getting from the
/create_success call (which we make by following a redirect from s3)
was including the ‘while(1);` at the beginning of the json response
(but only if the user has to do a full log in).

I tried interrupting the redirect (which we can do, just not with
background NSURLSessions) and setting the content-type headers, passing
the authorization, etc all to no avail. Stripping the ‘while(1);’ out
of the response by hand is the only solution I could come up with. It’s
not great but it works.

We only do the stripping if the parseData fails once then we check if
‘while(1);’ exists and try parsing again so this won’t effect every
successful call.