[MBL-7368] Error Mischief Managed (#302)
* Error Mischief Managed

All errors are funneled through the `ErrorReporter`. Canvas now
uses a new utility in `SoLazy` that categorizes errors. wether
theyre reportable to support, wether they should be logged to
crashlytics and wether they should be shown to the user at all

p.s. I also snuck in a fix that makes the `weight` on an
`AssignmentGroup` default to 0.0 if its not set.

* copy change

* Dont change the LegoSets propTypes

* Clean up error codes

* Fix EverythingBagel tests