MBL-9547 - Custom schemes in student and teacher (#1447)
* MBL-9547 - Custom schemes in student and teacher

This commit will make it possible using the canvas-teacher
and canvas-student schemes to navigate to their respective apps.
The only views in teacher that can be navigated to using this
method are views we have made work for deep linking. In student
the only views that work with deep linking currently are old
student views. None of our new RN views will work.

Test Plan:
 - Go to safari and type in canvas-teacher://<canvas-domain>.instructure.com
 - It should take you back to the dashboard
 - Go to safari and type in
   canvas-teacher://<canvas-domain>.instructure.com/courses/:courseID/discussion/:discussionID
 - It should modally show you the discussion
 - Go to safari and type in
   canvas-student://<canvas-domain>.instructure.com/courses/:courseID/assignments
 - It should take you to the student app assignment list

* Fix student app logging people out

* Fix flow error

* Fix teacher project