MBL-10048 - Fixes modal leak where modals caused some vcs to get left around (#1677)
* Fixes modal leak where modals caused some vcs to get left around

I think the big culprit was that we weren't dismissing modals prior
to restarting the app and I think UIKit was just not handling having
the root view controller pulled out from underneath it while a modal
was open. Also there was a bug with dismissAllModals where it wasn't
actually dismissing all the modals. This fix should ensure that all
modals are gone before any cleanup and restarting happens

* Don't try to dismiss if there are no modals