[MBL-10196] Don't munge endAt in calendar (#1809)
For all day events, we were setting an endAt previously to try to make up for
the core data predicate being exclusive on both ends. Since now it is inclusive
at the start, the setting the endAt date was causing it to appear on the next
day as well. I removed the old endAt munging since the predicate works right.

Test Plan:
- go to calendar
- verify all day events only appear on the correct day.
- verify events that happen at midnight appear only on the new day.
- whack any moles that pop up.