[MBL-10181] Include and indicate deleted comments (#1842)
Annotations delted by a teacher are not displayed at all. While this does not
match web, that is the current design for mobile.
Comments deleted by a teacher are displayed to the user that made them, along
with a red message about who deleted it and when. When the Comments view is
dismissed the deleted comments are removed and acknowledged, so they don't show
up again later.

Test Plan:
- As a student create some annotations and comments on annotations
- As a teacher delete a few of them
- As the student go back and view the submission
- verify deleted annotations do not show up
- open comments that should include deleted ones
- verify deleted comments are displayed with removed message
- close the comments view
- open the comments view again
- verify the deleted comments are gone