[MBL-10497] Reload details instead of trying to do intelligent diffing (#1913)
FetchedDetailsCollection was crashing on NSFetchedResultsController
change callbacks saying that the number of items were out of sync with
the data source. We saw this a lot with FetchedCollection and after
fighting it for months we decided to just do a hard `.reload` for all
changes. I think it's safe to that here for details as well (especially since
there are only 2 or three views that use FetchedDetailsCollection and
they don't need intelligent diffing).