Fix crash in Compose (#2151)
refs: none
affects: student, teacher
release note: Fixed a bug that would cause a crash when composing a message

test plan:
Check out the breadcrumbs in this bugsnag crash:
https://app.bugsnag.com/instructure-inc-1/student/errors/5b4bf7babb251300183e09ba?event_id=5b69f2590028cc4baa070000&i=em&m=fq

I think it's something to do with lodash's `throttle` but I can't be
sure. The breadcrumbs make it clear that it's composing from `Compose
Message` and obviously `this.scrollView` is now null. So this should fix
it.