Use help links from API (#2141)
refs: MBL-10964
affects: student, teacher
release note: none

test plan:
Until the api endpoint is merged, there will be a default help option
displayed which is to "Report a Problem" very similar to how it
currently works. The default help option is set when the request to get
the help links fails which will always be the case until the api
endpoint is merged. No error is thrown, it fails silently with the
defaults.

So, the test plan for now can be that "Report a Problem" is shown when
the user taps "Help" from global nav.

If you can get canvas running locally you can checkout my patchset with
the api endpoint and test it against that to see the help links from the
api.