Exhaust pagination using ttl and force flag (#2218)
* Exhaust pagination using ttl and force flag

[ignore-commit-lint]

`PaginatedUseCase` is an operation that will exhaust an
`APIRequestable`.

`TTLOperation` is an operation that will run an operation if the ttl has
expired or if `force` is true.

`GetCourses` is now a `PaginatedUseCase` wrapped in a `TTLOperation`.

Each page is written to the DatabaseStore as they are received. This is
good and bad.

* Good because the user will start seeing results as soon as the first
page is loaded.
* Bad because if we sort differently than the order of the results then the
table rows may jump around a bit which is annoying.

This is just one way to handle pagination. I'm open to exploring other
ideas.