Strip frameworks to fix submitting to App Store (#2256)
[ignore-commit-lint]

This is to workaround an [App Store submission bug](http://www.openradar.me/radar?id=6409498411401216)
when archiving universal binaries which we do as of Secrets.framework.

The bug is that the App Store rejects apps with frameworks built for simulator architectures.

This adds Realm's [strip-frameworks.sh](https://github.com/realm/realm-cocoa/blob/master/scripts/strip-frameworks.sh)
because it works well and strips all of the frameworks. PSPDFKit [has one](https://pspdfkit.com/guides/ios/current/getting-started/integrating-pspdfkit/#integrating-the-dynamic-framework)
too but it's specific to their frameworks.

I put it in CanvasCore for now. Core would be a better place but I
need it for Parent asap and Parent isn't linking against Core yet.