Fix ModuleListPresenter tests
[ignore-commit-lint]

The main issue here was that `UseCase.fetch` callbacks were being
completed in tests that did not initiate them.

- TestA kicks off a request for courses and a request for modules
- TestA waits until courses load then the test would finish
- TestB starts
- TestB mocks modules request
- TestB requests modules
- TestA's request for modules would finish
- **TestA's modules would get written to TestB's db**
- TestB's request for modules would finish
- TestB's modules would get written to TestB's db
- TestB would fail because too many modules were written

This is due to UseCases retaining themselves until the callbacks
complete and not having a way to cancel a UseCase..

We should _probably_ support canceling at some point but in the
meantime, ensuring that we write to the same database that initiated the
`fetch` fixes test issues since we re-create the db for every test.