Fix a crash on startup (#225)
[ignore-commit-lint]

I got a `BAD_ACCESS` error here. It originated from trying to set up
push notifications in `getNotificationPreferencesSetup`.

I'm not positive that this will fix the crash but it seems like we
should not strongly retain `self` here.

I tested areas of the app that use this ThreeLegit stuff and they
appear to still be working.