Workaround CourseNavigation header issue
The bad scrolling happens when the initial refresh flashes. By not reporting to be refreshing, the bad scroll never happens.
This does mean however that the spinner won't be accurately shown for the whole time it is refreshing.

refs: MBL-13034
affects: Student, Teacher
release note: Fixed an issue with Course navigation scrolling