Fix attaching from My Files
The native route handler lost the function props passed by the attachment picker. When there are props passed to an RN route from an RN screen, it has to be handled all in Helm to get the props passed through.

refs: MBL-13309
affects: Student
release note: Fixed an issue with attaching a file from My Files