Add prerequisite module UI test (#932)
* Add prerequisite module test

[run-nightly]

refs: MBL-13692
affects: student
release note: none

* Fix potential deadlock when viewing page module item

I think the problem must be that we try to invalidate the same cache right in a row.

* Prevent occassional deadlock on the main thread

I'm thinking that the deadlock could only happen in UI tests because the deadlock was on `contextsToLock[1].performAndWait` and `CFMessagePortSendRequest` in IPC.swift. But, I don't think we need all this manual locking. I'm sure Derrick had reasons but I'm not seeing any reason why we would need to use `performAndWait` for `mergeChanges` here.

The deadlock would occur when viewing a module item invalidated the refresh cache for modules and trigger a refresh of the API.

This the commit for where this was added initially: https://github.com/instructure/ios/pull/197/commits/ad63ca4836f67ae85b1b1a507ce59f91e00bce64