Handle page assignments
Page assignments are new to me. I don't think this should be a thing but the
web does display this page in the assignments list so /shrug.

The assignment object has no reference to the page so we have no way of knowing to
direct to the page or to load the page body so we load a weird assignment details
screen for assignment pages /shrug.

refs: MBL-13880
affects: student
release note: Fixed a bug that would prevent the assignments list from loading after unlocking a mastery path

Test plan:
* Log in as the user in [the ticket](https://instructure.atlassian.net/browse/MBL-13880) and follow the steps to get to the Assignments page
* The Assignments page should load