Fix loading files that are locked
refs: MBL-13971
affects: student
release note: Fixed a bug that would prevent the To Do list from loading

When a file has locked_for_user=true the url is an empty string. Instead
of making the URL nil in that case the url is set to the current
session's baseURL.

This way we can keep File.url non-optional and avoid writing a custom
decode initializer.

If we'd rather File.url be nil in the case I can make that happen.

With this, we should probably get in the habit of checking
`lockedForUser` when presenting the file url.