Add calendar deep link
test plan:
twilson > Assignment Overrides has a link to the calendar in the syllabus body.
Clicking that should open the new calendar if the flag is enabled, same as other links there work.
Deep links to /calendar should work as well, opening a modal as is typical.

refs: MBL-13906
affects: Parent
release note: none