Fix 'Show Grades' option
I had to update the CacheManager to not blow away everything in the
`Bundle.main.appGroupID()` container because that includes shared
UserDefaults which is where `SessionDefault`s are stored including this
flag.

Instead, CacheManager only deletes the newly added `caches/` directory
in the App Group. Core Data files are now stored in that caches
directory instead of at the root.

refs: MBL-14332
affects: student
release note: Fixed a bug that would cause the 'Show Grades' option to reset

Test plan:
- Checkout the changes locally
- Build and run the app
- From the global nav, toggle 'Show Grades' to On
- Change the build number of Student in Xcode
- Build and run again
- Grades on dashboard should still be shown