Fix loading some group discussions
The API returns all course groups when we ask for them, even if the
current user isn't in the group. And all the group topics are included
in the discussion, even those that are not for the current user.

So for child group topics we should only request the groups for the
current user so that we can accurately narrow down which child topic to
show.

refs: MBL-14543
affects: student
release note: Fixed loading discussions

Test plan:
- Install student from master or prod
- Masquerade as users in ticket
- View the linked discussions
- They should load