Add extended late policy status. (#2171)
refs: MBL-16064
affects: Student, Teacher, Parent
release note: none

test plan: This change adds the extended enumeration option to LatePolicyStatus to make sure that when this is available on the API the app won't fail parsing it. Since we don't use this field in the app there's no way to manually test it.